♻️ 1E0EHD backend: split cloud start refresh
