+++ .agentplane/tasks/202605191451-1E0EHD/pr/meta.json
@@ -3,8 +3,8 @@
   "branch": "task/202605191451-1E0EHD/daily-cloud-start-pull",
   "created_at": "2026-05-19T14:52:45.909Z",
   "head_sha": "bf213af3cbf89aec9404e5c359ab39e7ab54de90",
-  "last_verified_at": "2026-05-19T15:08:21.753Z",
-  "last_verified_sha": "81a3ed59446b0e0fbdf663711e789c53ec915369",
+  "last_verified_at": "2026-05-19T15:36:20.083Z",
+  "last_verified_sha": "bf213af3cbf89aec9404e5c359ab39e7ab54de90",
   "pr_number": 3938,
   "pr_url": "https://github.com/basilisk-labs/agentplane/pull/3938",
   "schema_version": 1,
